# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CSE 341 Project 2 API",
        "description": "REST API performing full CRUD operations on the contacts, users, colleges, courses, and instructors collections in MongoDB. All routes include data validation and error handling, and return proper HTTP status codes. Every POST, PUT, and DELETE route requires an authenticated GitHub session and returns 401 otherwise. To try a protected route in Swagger UI, open /login in a browser and sign in with the same host, then reload this page.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://web-servises-cse-341.onrender.com/",
            "description": "Production (Render)"
        },
        {
            "url": "http://localhost:3000",
            "description": "Local development"
        }
    ],
...
    "paths": {
...
        "/instructors": {
            "get": {
                "tags": ["Instructors"],
                "operationId": "getInstructors",
                "summary": "Get all instructors",
                "description": "Returns a list of every instructor in the instructors collection.",
                "responses": {
                    "200": {
                        "description": "A list of instructors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Instructor"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "An unexpected error occurred"
                    }
                }
            },
            "post": {
                "tags": ["Instructors"],
                "operationId": "createInstructor",
                "security": [{ "githubOAuth": [] }],
                "summary": "Create an instructor",
                "description": "Inserts a new instructor into the instructors collection. courseId must match an existing course.",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Instructor"
                            },
                            "examples": {
                                "instructor": {
                                    "value": {
                                        "firstName": "Grace",
                                        "lastName": "Nabirye",
                                        "email": "g.nabirye@mak.ac.ug",
                                        "phone": "+256-414-422-101",
                                        "title": "professor",
                                        "courseId": "64f1e2d3a5b6c7d8e9f0a1b3",
                                        "faculty": "Faculty of Technology",
                                        "office": "Tech B14",
                                        "hireDate": "2009-02-01"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Instructor created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "example": "64f1e2d3a5b6c7d8e9f0a1b4"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Missing required fields, invalid data, or unknown courseId"
                    },
                    "401": {
                        "description": "Not authenticated; log in with GitHub first"
                    },
                    "500": {
                        "description": "An unexpected error occurred"
                    }
                }
            }
        },
        "/instructors/{id}": {
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "required": true,
                    "description": "The id of the instructor",
                    "schema": {
                        "type": "string"
                    }
                }
            ],
            "get": {
                "tags": ["Instructors"],
                "operationId": "getInstructorById",
                "summary": "Get an instructor by id",
                "description": "Returns a single instructor matching the given id.",
                "responses": {
                    "200": {
                        "description": "The requested instructor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Instructor"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid instructor id"
                    },
                    "404": {
                        "description": "Instructor not found"
                    },
                    "500": {
                        "description": "An unexpected error occurred"
                    }
                }
            },
            "put": {
                "tags": ["Instructors"],
                "operationId": "updateInstructor",
                "security": [{ "githubOAuth": [] }],
                "summary": "Update an instructor by id",
                "description": "Replaces the fields of the instructor matching the given id. All fields are required.",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Instructor"
                            },
                            "examples": {
                                "instructor": {
                                    "value": {
                                        "firstName": "Grace",
                                        "lastName": "Nabirye",
                                        "email": "g.nabirye.updated@mak.ac.ug",
                                        "phone": "+256-414-422-101",
                                        "title": "senior lecturer",
                                        "courseId": "64f1e2d3a5b6c7d8e9f0a1b3",
                                        "faculty": "Faculty of Technology",
                                        "office": "Tech B21",
                                        "hireDate": "2009-02-01"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Instructor updated"
                    },
                    "400": {
                        "description": "Invalid instructor id, missing fields, or unknown courseId"
                    },
                    "401": {
                        "description": "Not authenticated; log in with GitHub first"
                    },
                    "404": {
                        "description": "Instructor not found"
                    },
                    "500": {
                        "description": "An unexpected error occurred"
                    }
                }
            },
            "delete": {
                "tags": ["Instructors"],
                "operationId": "deleteInstructor",
                "security": [{ "githubOAuth": [] }],
                "summary": "Delete an instructor by id",
                "description": "Deletes the instructor matching the given id.",
                "responses": {
                    "204": {
                        "description": "Instructor deleted"
                    },
                    "400": {
                        "description": "Invalid instructor id"
                    },
                    "401": {
                        "description": "Not authenticated; log in with GitHub first"
                    },
                    "404": {
                        "description": "Instructor not found"
                    },
                    "500": {
                        "description": "An unexpected error occurred"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "githubOAuth": {
                "type": "oauth2",
                "flows": {
                    "authorizationCode": {
                        "authorizationUrl": "https://github.com/login/oauth/authorize",
                        "tokenUrl": "https://github.com/login/oauth/access_token",
                        "scopes": {
                            "user:email": "Read access to your email address"
                        }
                    }
                }
            }
        },
        "schemas": {
...
            "Instructor": {
                "type": "object",
                "required": ["firstName", "lastName", "email", "phone", "title", "courseId", "faculty", "office", "hireDate"],
                "properties": {
                    "firstName": {
                        "type": "string",
                        "example": "Grace"
                    },
                    "lastName": {
                        "type": "string",
                        "example": "Nabirye"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "g.nabirye@mak.ac.ug"
                    },
                    "phone": {
                        "type": "string",
                        "example": "+256-414-422-101"
                    },
                    "title": {
                        "type": "string",
                        "enum": ["professor", "associate professor", "senior lecturer", "lecturer", "assistant lecturer", "teaching assistant"],
                        "example": "professor"
                    },
                    "courseId": {
                        "type": "string",
                        "description": "The _id of the course this instructor teaches. Must match an existing course.",
                        "example": "64f1e2d3a5b6c7d8e9f0a1b3"
                    },
                    "faculty": {
                        "type": "string",
                        "example": "Faculty of Technology"
                    },
                    "office": {
                        "type": "string",
                        "description": "Office location or room number.",
                        "example": "Tech B14"
                    },
                    "hireDate": {
                        "type": "string",
                        "description": "Date the instructor was hired, in YYYY-MM-DD format.",
                        "example": "2009-02-01"
                    }
                }
            }
        }
    }
}